% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cleanWS}
\alias{cleanWS}
\title{Cleans the workspace including unloading of libraries.}
\usage{
cleanWS()
}
\description{
This is usefull for verifying that all calls to \code{library} are correctly placed in the R code.
}
\details{
The function
\enumerate{
  \item removes all objects
  \item Detachs all packages except: package:stats","package:graphics","package:grDevices","package:utils","package:datasets","package:methods","package:base"
  \item Unloads all namespaces except: "stats","graphics","grDevices","utils","datasets","base"
}
}

