% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_stamp.R
\name{git_stamp}
\alias{git_stamp}
\title{Get git stamp (commit and branch) for a repository}
\usage{
git_stamp(repo = getwd())
}
\arguments{
\item{repo}{Git repo directory. If unspecified, then the current working
directory is used.}
}
\value{
character vector with latest commit and branch
}
\description{
The function returns the latest git commit and branch for the repo specified
 in \code{repo} or the current working directory if unspecified.
 Git is needed for the command to run.
 The functions makes it possible to include the latest git commit and branch
 in a run to be able to know exactly which code where used.
}
