\name{PackInstallCheck}
\alias{PackInstallCheck}
\title{Subset a vectors of packages and package versions to those that are that are not already installed.}
\usage{
  PackInstallCheck(pkgs = NULL, versions = NULL,
    lib = NULL)
}
\description{
  Subset a vectors of packages and package versions to
  those that are that are not already installed.
}
\keyword{internal}

