% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repmis.R
\name{Repmis}
\alias{Repmis}
\title{Replace missing values}
\usage{
Repmis(x, fun = median, verbose = FALSE)
}
\arguments{
\item{x}{Numeric Vector or Vector}
}
\value{
A vector containing NA values replaced
}
\description{
Replace missing values
}
\examples{
Repmis(x, fun = mean, verbose = TRUE)
}
