% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scan_https.R
\name{scan_https}
\alias{scan_https}
\title{Read a character text file from a secure (https) site into R as a single
object.}
\source{
Originally based on source_url from the Hadley Wickham's devtools
package.
}
\usage{
scan_https(url, sha1 = NULL)
}
\arguments{
\item{url}{The files's URL.}

\item{sha1}{Character string of the file's SHA-1 hash, generated by
\code{source_data}. Note if you are using data stored using Git, this is not
the file's commit SHA-1 hash.}
}
\value{
a charcter object of length 1
}
\description{
Read a character text file from a secure (https) site into R as a single
object.
}

