% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_valid_wd}
\alias{set_valid_wd}
\title{Sets valid working directory from vector of possible directories}
\usage{
set_valid_wd(possible)
}
\arguments{
\item{possible}{character vector of possible working directores}
}
\description{
Sets valid working directory from vector of possible directories
}
\details{
Sets the working directory to the first valid directory from a
list of possible directories.
}
\examples{
\dontrun{
set_valid_wd(c('examples/directory1', 'anotherExample/directory2'))
}

}
