% Generated by roxygen2 (4.0.1): do not edit by hand
\name{source_RData}
\alias{source_RData}
\title{Load plain-text data from a URL (either http or https)}
\source{
Based on source_url from the Hadley Wickham's devtools package.
}
\usage{
source_RData(url, sha1 = NULL, cache = FALSE, clearCache = FALSE,
  envir = parent.frame(), ...)
}
\arguments{
\item{url}{The plain-text formatted data's URL.}

\item{sha1}{Character string of the file's SHA-1 hash, generated by
\code{source_RData}. Note if you are using data stored using Git, this is not
the file's commit SHA-1 hash.}

\item{cache}{logical. Whether or not to cache the data so that it is not
downloaded every time the function is called.}

\item{clearCache}{logical. Whether or not to clear the downloaded data from
the cache.}

\item{envir}{the environment where the data should be loaded.}

\item{...}{additional arguments passed to \code{\link{load}}.}
}
\value{
data stored in the RData file
}
\description{
\code{source_RData} loads R data at a URL (both http
and https) into R.
}
\details{
Loads Rdata from a URL. Works with
both HTTP and HTTPS sites. Note: the URL you give for the \code{url} argument
must be for the RAW version of the file. The function should work to download
plain-text data from any secure URL (https), though I have not verified this.

From the \code{source_url} documentation: "If a SHA-1 hash is specified with
the \code{sha1} argument, then this function will check the SHA-1 hash of the
downloaded file to make sure it matches the expected value, and throw an error
if it does not match. If the SHA-1 hash is not specified, it will print a
message displaying the hash of the downloaded file. The purpose of this is to
 improve security when running remotely-hosted code; if you have a hash of
the file, you can be sure that it has not changed."
}
\seealso{
\link{httr} and \code{\link{read.table}}
}

